refactor: move RegisterModal to LoginModal folder
