+++ src/app/layout.tsx
@@ -5,10 +5,10 @@ import { GoogleAnalytics } from '@next/third-parties/google';
 import { type Metadata } from 'next';
 import { Bai_Jamjuree, Inter } from 'next/font/google';
 import { CheckRefreshToken } from '@src/components/auth/CheckRefreshToken';
-import { RegisterModalProvider } from '@src/components/global/RegisterModalProvider';
 import { SnackbarProvider } from '@src/components/global/Snackbar';
 import { TRPCReactProvider } from '@src/trpc/react';
 import ClientLocalizationProvider from '@src/utils/localization';
+import { RegisterModalProvider } from '@src/utils/LoginModal/provider';
 import theme from '@src/utils/theme';
 
 const inter = Inter({
