+++ src/components/events/directory/filter/panels/Filters.tsx
@@ -5,10 +5,10 @@ import ToggleButton from '@mui/material/ToggleButton';
 import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
 import { memo } from 'react';
 import Panel from '@src/components/common/Panel';
-import { useRegisterModal } from '@src/components/global/RegisterModalProvider';
 import { setSnackbar } from '@src/components/global/Snackbar';
 import { authClient } from '@src/utils/auth-client';
 import { EventFiltersSchema } from '@src/utils/eventFilter';
+import { useRegisterModal } from '@src/utils/LoginModal/provider';
 import { FilterPanelProps, panelProps, setEventsParams } from '../utils';
 
 export type FiltersPanelFields = Pick<
